requirements.txt
pyTelegramBotAPI==4.4.0
psycopg2==2.9.3
python-dotenv==0.20.0